yt_summarizer/requirements.txt
langchain==0.0.208
deeplake==3.6.5
openai==0.27.8
tiktoken==0.4.0
requests
yt_dlp
openai-whisper
